Update project to improve code structure and fix linting errors
Refactors the demo and theme providers into separate hooks and contexts, introduces new context files, and adjusts ESLint rules to enforce stricter component export requirements.
