+++ client/src/contexts/theme-context.ts
@@ -1,0 +1,15 @@
+import { createContext } from "react";
+
+export type Theme = "dark" | "light" | "system";
+
+export interface ThemeProviderState {
+  theme: Theme;
+  setTheme: (theme: Theme) => void;
+}
+
+const initialState: ThemeProviderState = {
+  theme: "system",
+  setTheme: () => null,
+};
+
+export const ThemeProviderContext = createContext<ThemeProviderState>(initialState);
