+++ client/src/providers/theme-provider.tsx
@@ -1,1 +1,2 @@
-export { ThemeProvider, useTheme } from "@/hooks/use-theme";
+export { ThemeProvider } from "@/hooks/use-theme";
+export { useTheme } from "@/hooks/use-theme-hook";
